Split responsibilities around CLI into different classes
