+++ Source/DafnyDriver/Commands/MeasureComplexityCommand.cs
@@ -37,9 +37,9 @@ static class MeasureComplexityCommand {
     foreach (var option in Options) {
       result.AddOption(option);
     }
-    DafnyCli.SetHandlerUsingDafnyOptionsContinuation(result, (options, _) => {
+    DafnyNewCli.SetHandlerUsingDafnyOptionsContinuation(result, (options, _) => {
       options.Compile = false;
-      return CompilerDriver.RunCompiler(options);
+      return LegacyCliCompilation.Run(options);
     });
     return result;
   }
