+++ Source/DafnyDriver/Commands/TranslateCommand.cs
@@ -40,12 +40,12 @@ static class TranslateCommand {
         subCommand.AddGlobalOption(option);
       }
 
-      DafnyCli.SetHandlerUsingDafnyOptionsContinuation(subCommand, async (options, context) => {
+      DafnyNewCli.SetHandlerUsingDafnyOptionsContinuation(subCommand, async (options, context) => {
         options.Compile = false;
         var noVerify = options.Get(BoogieOptionBag.NoVerify);
         options.CompilerName = subCommand.Name;
         options.SpillTargetCode = noVerify ? 3U : 2U;
-        var continueCliWithOptions = await CompilerDriver.RunCompiler(options);
+        var continueCliWithOptions = await LegacyCliCompilation.Run(options);
         return continueCliWithOptions;
       });
     }
