+++ Source/DafnyPipeline.Test/RelativePaths.cs
@@ -16,7 +16,7 @@ namespace DafnyPipeline.Test {
 
     [Fact]
     public void Test() {
-      Assert.Equal(0, DafnyCli.MainWithWriters(output, output,
+      Assert.Equal(0, DafnyBackwardsCompatibleCli.MainWithWriters(output, output,
         TextReader.Null, new[] { "/spillTargetCode:3", "warnings-as-errors.dfy" }));
     }
